refactor(azure-auth): move auth-mode cache from model_client to azure_auth
Pure mechanical move; no behavior change. The auth-mode cache
(`_AZURE_AUTH_MODE`, `_AZURE_AAD_DEP_MISSING`,
`refresh_azure_auth_mode`, `_get_azure_auth_mode`,
`log_resolved_azure_auth_mode`) now lives in azure_auth.py next to the
existing token-provider cache, so a single module owns Azure auth state.

model_client.py reaches across via `azure_auth.<symbol>` at the two
call sites (`_maybe_inject_azure_aad_token`, `_classify_llm_error`).
runner.py / init/_command.py / cli.py import refresh/log helpers from
the new home directly.

Tests rebound minimally (patch targets repointed at the owner module).
Consolidation of the cache-reset boilerplate via a helper follows in the
next commit.
