+++ assert_ai/init/_command.py
@@ -140,7 +140,7 @@ def init(
     # environment so the init LLM call respects the dotenv-supplied
     # ``ASSERT_AZURE_USE_AAD`` / ``AZURE_API_KEY`` rather than whichever
     # state was frozen at module import.
-    from assert_ai.core.model_client import log_resolved_azure_auth_mode, refresh_azure_auth_mode
+    from assert_ai.core.azure_auth import log_resolved_azure_auth_mode, refresh_azure_auth_mode
 
     refresh_azure_auth_mode(force=True)
     # Emit the resolved auth-mode line AFTER refresh so it reflects the
